Updates to ENSManager to allow manager offchain signature to be passed Upgrade ENS contracts to 0.6.12 and cleanup Managed contract duplication which is no longer necessary
