--- contracts/infrastructure/base/ManagedV2.sol
@@ -1,64 +1,0 @@
-// Copyright (C) 2018  Argent Labs Ltd. <https://argent.xyz>
-
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-
-// You should have received a copy of the GNU General Public License
-// along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-// SPDX-License-Identifier: GPL-3.0-only
-pragma solidity ^0.6.12;
-
-import "./Owned.sol";
-
-/**
- * @title Managed
- * @notice Basic contract that defines a set of managers. Only the owner can add/remove managers.
- * infrastructure/base/ManagedV2.sol differs from infrastructure_0.5/base/Managed.sol in being solidity 0.6 compatible
- * @author Julien Niset, Olivier VDB - <julien@argent.xyz>, <olivier@argent.xyz>
- */
-contract ManagedV2 is Owned {
-
-    // The managers
-    mapping (address => bool) public managers;
-
-    /**
-     * @notice Throws if the sender is not a manager.
-     */
-    modifier onlyManager {
-        require(managers[msg.sender] == true, "M: Must be manager");
-        _;
-    }
-
-    event ManagerAdded(address indexed _manager);
-    event ManagerRevoked(address indexed _manager);
-
-    /**
-    * @notice Adds a manager.
-    * @param _manager The address of the manager.
-    */
-    function addManager(address _manager) external onlyOwner {
-        require(_manager != address(0), "M: Address must not be null");
-        if (managers[_manager] == false) {
-            managers[_manager] = true;
-            emit ManagerAdded(_manager);
-        }
-    }
-
-    /**
-    * @notice Revokes a manager.
-    * @param _manager The address of the manager.
-    */
-    function revokeManager(address _manager) external virtual onlyOwner {
-        require(managers[_manager] == true, "M: Target must be an existing manager");
-        delete managers[_manager];
-        emit ManagerRevoked(_manager);
-    }
-}
